pyproject.toml
[tool.poetry]
name = "c-facade"
version = "0.1.19"
description = ""
authors = ["lwh <whl@provenance.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.3"
boto3 = "^1.34.19"
fastapi = "^0.108.0"
fastapi-camelcase = "^2.0.0"
google-cloud-storage = "^2.14.0"
google-cloud-vision = "^3.7.2"
gunicorn = "^21.2.0"
httpx = "^0.26.0"
jinja2 = "^3.1.2"
playwright = "^1.50.0"
playwright-stealth = "^1.0.6"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
pymilvus = "^2.5.0"
pymongo = "^4.10.1"
python = "^3.9"
python-json-logger = "^2.0.7"
redis = "^5.2.1"
selenium = "^4.19.0"
uvicorn = "^0.25.0"
webdriver-manager = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
